Move parser inside analyzer module
