+++ analyzer/config.py
@@ -2,3 +2,6 @@
 CIRCLE_CI = "CircleCI"
 GH_ACTIONS = "Github Actions"
 TRAVIS_CI = "TravisCI"
+
+GH_PROVIDERS = ['github', 'gh']
+PROVIDERS = [*GH_PROVIDERS]
